Still figure out how to display a list of alphabet before they going to breed
