+++ project-two/src/App.js
@@ -1,6 +1,6 @@
 import './styles/App.css';
 import { Routes, Route } from 'react-router-dom';
-import DogDetails from './Components/DogDetails'
+import BreedPhotos from './Components/BreedPhotos'
 import Navbar from './Components/Navbar'
 import RandomDog from './Components/RandomDog';
 import DogList from './Components/DogList';
@@ -16,7 +16,7 @@ function App() {
       <Routes>
         <Route path="/" element={ <RandomDog  />  } />
         <Route path="/list" element={<DogList />} />
-        <Route path="/list/:nameDog" element={ <DogDetails /> } />
+        <Route path="/list/:nameDog" element={ <BreedPhotos /> } />
       </Routes>
     </main>
     <footer>
